+++ rust/examples/05.rs
@@ -1,13 +1,13 @@
 use advent_2019::intcode::{compute, parse_input};
 use advent_2019::Result;
 
-fn part1() -> Result<i32> {
+fn part1() -> Result<i64> {
     let program = parse_input("../inputs/05-input.txt")?;
     let outputs = compute(program, vec![1])?;
     Ok(*outputs.last().ok_or("No outputs")?)
 }
 
-fn part2() -> Result<i32> {
+fn part2() -> Result<i64> {
     let program = parse_input("../inputs/05-input.txt")?;
     let outputs = compute(program, vec![5])?;
     Ok(*outputs.last().ok_or("No outputs")?)
